requirements.txt
tensorflow-gpu >= 1.2.0
numpy
scipy
matplotlib
seaborn
